+++ spec/controllers/api/courses_controller_spec.rb
@@ -157,6 +157,11 @@ describe Api::CoursesController do
       put :reset_autotest_connection, params: { id: course.id }
       expect(response).to have_http_status(:forbidden)
     end
+
+    it 'should fail to authenticate a PUT refresh_autotest_schema request' do
+      put :refresh_autotest_schema, params: { id: course.id }
+      expect(response).to have_http_status(:forbidden)
+    end
   end
 
   context 'An instructor user in the course' do
@@ -195,6 +200,11 @@ describe Api::CoursesController do
 
     it 'should fail to authenticate a PUT update request' do
       get :update, params: { id: course.id }
+      expect(response).to have_http_status(:forbidden)
+    end
+
+    it 'should fail to authenticate a PUT refresh_autotest_schema request' do
+      put :refresh_autotest_schema, params: { id: course.id }
       expect(response).to have_http_status(:forbidden)
     end
 
@@ -526,6 +536,43 @@ describe Api::CoursesController do
 
         it 'should fail when an error is raised' do
           allow(AutotestResetUrlJob).to receive(:perform_now).and_raise(StandardError)
+          subject
+          expect(response).to have_http_status(:internal_server_error)
+        end
+      end
+    end
+
+    describe '#refresh_autotest_schema' do
+      subject { put :refresh_autotest_schema, params: { id: course.id } }
+
+      context 'there is no autotest_setting set' do
+        it 'should return unprocessable_entity' do
+          subject
+          expect(response).to have_http_status(:unprocessable_content)
+        end
+      end
+
+      context 'there is an autotest_setting' do
+        include_context 'course with an autotest setting'
+        let(:new_schema) { { 'title' => 'new' }.to_json }
+
+        it 'should store the schema fetched from the autotester' do
+          allow(controller).to receive(:get_schema).and_return(new_schema)
+          subject
+          expect(response).to have_http_status(:ok)
+          expect(course.autotest_setting.reload.schema).to eq new_schema
+        end
+
+        it 'should return unprocessable_entity when the record is invalid' do
+          allow(controller).to receive(:get_schema).and_return(new_schema)
+          allow_any_instance_of(AutotestSetting).to receive(:update!)
+            .and_raise(ActiveRecord::RecordInvalid.new(AutotestSetting.new))
+          subject
+          expect(response).to have_http_status(:unprocessable_content)
+        end
+
+        it 'should fail when the autotester cannot be reached' do
+          allow(controller).to receive(:get_schema).and_raise(StandardError, 'down')
           subject
           expect(response).to have_http_status(:internal_server_error)
         end